tests/requirements.txt
jingo
celery
django-celery
South
